Uniq/try_on.py
import os
import torch
import torch.nn as nn
from torchvision.utils import save_image
from networks import ConditionGenerator, load_checkpoint
import argparse
from PIL import Image
import torchvision.transforms as transforms
import numpy as np

def get_opt():
    parser = argparse.ArgumentParser()
    parser.add_argument("--person_image", required=True, help="Path to person image")
    parser.add_argument("--clothing_image", required=True, help="Path to clothing image")
    parser.add_argument("--output_dir", default="./user_output", help="Output directory")
    parser.add_argument("--gpu_ids", default="0")
    parser.add_argument("--tocg_checkpoint", default="checkpoints/mtviton.pth")
    parser.add_argument("--gen_checkpoint", default="checkpoints/gen.pth")
    return parser.parse_args()

def preprocess_image(image_path, size=(768, 1024)):
    transform = transforms.Compose([
        transforms.Resize(size),
        transforms.ToTensor(),
        transforms.Normalize((0.5,), (0.5,))
    ])
    image = Image.open(image_path).convert('RGB')
    return transform(image).unsqueeze(0)

def main():
    opt = get_opt()
    os.makedirs(opt.output_dir, exist_ok=True)
    
    # Set device
    device = torch.device(f"cuda:{opt.gpu_ids}" if torch.cuda.is_available() else "cpu")
    
    # Load models
    input1_nc = 4  # cloth + cloth-mask
    input2_nc = 13 + 3  # parse_agnostic + densepose
    tocg = ConditionGenerator(opt, input1_nc=input1_nc, input2_nc=input2_nc, output_nc=13, ngf=96, norm_layer=nn.BatchNorm2d)
    
    # Load checkpoints
    load_checkpoint(tocg, opt.tocg_checkpoint)
    tocg.to(device)
    tocg.eval()
    
    # Load and preprocess images
    person_img = preprocess_image(opt.person_image)
    clothing_img = preprocess_image(opt.clothing_image)
    
    # TODO: Add preprocessing steps for:
    # 1. OpenPose for person pose
    # 2. Human parsing
    # 3. DensePose
    # 4. Clothing mask
    # 5. Parse agnostic
    
    print("Note: This is a basic setup. To get full functionality, you need to:")
    print("1. Install OpenPose for pose detection")
    print("2. Install CIHP_PGN for human parsing")
    print("3. Install Detectron2 for DensePose")
    print("4. Use a background removal tool for clothing masks")
    print("5. Generate parse agnostic images")
    print("\nFor now, you can use the existing dataset images as examples:")
    print("Person images are in: data/train/image/")
    print("Clothing images are in: data/train/cloth/")

if __name__ == "__main__":
    main() 