Uniq/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
torchgeometry>=0.1.2
Pillow>=8.0.0
tqdm>=4.50.0
tensorboardX>=2.0
scikit-image>=0.17.2
scipy>=1.6.0
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0 